analyzer/consumer/requirements.txt
cffi==1.17.1
cryptography==45.0.4
greenlet==3.2.3
kafka-python==2.2.13
pycparser==2.22
PyMySQL==1.1.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
